FE - CreateMatch: Refacor react-router + Add to navbar
